Refactor OAuth Broker service and enhance functionality
- Introduced a new OAuth Broker service structure, consolidating functionality into a dedicated package.
- Added support for tool policy management, including CRUD operations for tool policies.
- Updated existing OAuth Broker methods to improve clarity and maintainability.
- Enhanced datetime handling to use timezone-aware objects across the service.
- Removed deprecated OAuth Broker service file, streamlining the codebase.
- Updated integration and unit tests to reflect changes in the OAuth Broker service structure and functionality.
