+++ kamiwaza_sdk/services/models/base.py
@@ -1,6 +1,5 @@
-from typing import List, Optional, Union, Dict, Any
+from typing import List, Union
 from uuid import UUID
-from ...exceptions import APIError
 from ...schemas.models.model import Model, CreateModel
 from ...schemas.guide import ModelGuide
 from ...utils.quant_manager import QuantizationManager
