+++ kamiwaza_sdk/services/models/search.py
@@ -1,5 +1,4 @@
-from typing import List, Optional, Dict, Any, Union, Set
-from uuid import UUID
+from typing import List, Optional, Any
 from ...schemas.models.model import Model
 from ...schemas.models.model_search import ModelSearchRequest, ModelSearchResponse, HubModelFileSearch
